requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
fastapi==0.135.2
h11==0.16.0
idna==3.11
lgpio==0.2.2.0
mfrc522==0.0.7
ndeflib==0.3.3
pi-rc522==2.3.0
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.33.1
rpi-lgpio==0.6
spidev==3.5
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
